reorganize provider package by concern
Split the monolithic provider.go into files named by their responsibility:

- registry.go: Provider interface, types, registration, lookup, DisplayName
- credentials.go: credential discovery, reuse policy, status queries
- auth.go: auth flow types (AuthFlow, DeviceAuthFlow, ManualKeyAuthFlow)
- validation.go: input validators for auth flows
- status.go: statuspage fetching

The old layout had provider.go as a grab-bag and credentials_policy.go as
a fragment. The new layout gives each file a name that tells you what's in it.
