+++ internal/provider/registry.go
@@ -99,70 +99,6 @@ func AvailableIDs(cfg config.Config) []string {
 	return ids
 }
 
-// FindCredential checks for credentials for the given provider in
-// vibeusage storage, provider CLI paths, and environment variables.
-// Returns (found, source, path).
-func FindCredential(providerID string) (bool, string, string) {
-	p, ok := Get(providerID)
-	if !ok {
-		return false, "", ""
-	}
-	info := p.CredentialSources()
-	return config.FindProviderCredential(providerID, info.CLIPaths, info.EnvVars)
-}
-
-// CheckCredentials reports whether the given provider has credentials
-// and where they came from.
-func CheckCredentials(providerID string) (bool, string) {
-	found, source, _ := FindCredential(providerID)
-	if found {
-		return true, source
-	}
-
-	if !ExternalCredentialReuseEnabled() {
-		return false, ""
-	}
-
-	p, ok := Get(providerID)
-	if !ok {
-		return false, ""
-	}
-
-	for _, s := range p.FetchStrategies() {
-		if s.IsAvailable() {
-			return true, "provider_cli"
-		}
-	}
-
-	return false, ""
-}
-
-// GetAllCredentialStatus returns the credential status for every
-// registered provider.
-func GetAllCredentialStatus() map[string]config.CredentialStatus {
-	all := All()
-	status := make(map[string]config.CredentialStatus, len(all))
-	for id := range all {
-		hasCreds, source := CheckCredentials(id)
-		status[id] = config.CredentialStatus{
-			HasCredentials: hasCreds,
-			Source:         source,
-		}
-	}
-	return status
-}
-
-// IsFirstRun returns true if no registered provider has credentials.
-func IsFirstRun() bool {
-	for id := range All() {
-		hasCreds, _ := CheckCredentials(id)
-		if hasCreds {
-			return false
-		}
-	}
-	return true
-}
-
 // DisplayName returns the human-readable display name for the given
 // provider ID by looking it up in the registry. If the ID is not
 // registered, it returns the ID itself as a fallback.
@@ -172,17 +108,4 @@ func DisplayName(id string) string {
 		return id
 	}
 	return p.Meta().Name
-}
-
-// CountConfigured returns the number of registered providers that
-// have credentials.
-func CountConfigured() int {
-	count := 0
-	for id := range All() {
-		hasCreds, _ := CheckCredentials(id)
-		if hasCreds {
-			count++
-		}
-	}
-	return count
 }
