+++ internal/provider/registry_test.go
